Add addNew and bookmanagement function
