+++ index.js
@@ -1,4 +1,8 @@
-const cForm = document.querySelector('form');
+
+import { Bookmanagement } from "./modules/bookmanagement.js";
+import {addNew} from "./modules/addNew.js";
+
+const addForm = document.querySelector('form');
 const title = document.querySelector('#title');
 const author = document.querySelector('#author');
 
@@ -6,53 +10,9 @@ if (!JSON.parse(localStorage.getItem('books'))) {
   localStorage.setItem("books", JSON.stringify([]));
 }
 
-class Bookmanagement {
-
-  static add(title, author) {
-    let restoredData = JSON.parse(localStorage.getItem('books'));
-    let aBook = {
-      title: title,
-      author: author,
-      id: `${title}${author}`
-    }
-    restoredData.push(aBook);
-    localStorage.setItem("books", JSON.stringify(restoredData));
-  }
-
-  static display() {
-    let restoredData = JSON.parse(localStorage.getItem('books'));
-    const display = document.querySelector('#book-list');
-    restoredData.forEach(element => {
-      display.innerHTML += `
-      <div class="single-book">
-      <p>${element.title} and ${element.author}</p>
-      <button class="delete-button"  id="${element.id}" >REMOVE</button>
-      </div>`
-
-    })
-  }
-
-  static delete(e) {
-    let restoredData = JSON.parse(localStorage.getItem('books'));
-    restoredData = restoredData.filter((el) => {
-      if (el.id == e) {
-        return false;
-      } else {
-        return true;
-      }
-    })
-    localStorage.setItem("books", JSON.stringify(restoredData));
-    location.reload();
-  }
-}
-
 Bookmanagement.display();
 
-cForm.addEventListener('submit', (e) => {
-  e.preventDefault();
-  Bookmanagement.add(title.value, author.value);
-  location.reload();
-})
+addForm.addEventListener('submit', addNew);
 
 let delBtn = document.querySelectorAll('.delete-button');
 delBtn.forEach((button) => {
