+++ modules/addNew.js
@@ -1,0 +1,6 @@
+import { Bookmanagement } from "./bookmanagement.js";
+export function addNew(e) {
+  e.preventDefault();
+  Bookmanagement.add(title.value, author.value);
+  location.reload();
+};
